Fix update employee

+++ src/main/java/academy/softserve/museum/util/PathParser.java
@@ -3,8 +3,8 @@ package academy.softserve.museum.util;
 public class PathParser {
     private PathParser() {}
 
-    public static int getPathVariable(String path) {
+    public static Long getPathVariable(String path) {
         String firstLine = path.split("\n")[0].replaceAll("/", "");
-        return Integer.parseInt(firstLine);
+        return Long.parseLong(firstLine);
     }
 }
